requirements.txt
requests
PyCryptodome
colorama

